Add files via upload
So this is the data from https://ourworldindata.org/food-supply with two notebooks, one that uses the entire dataset and the other that is the EU and the US. After looking at visualizations for just the EU and US the diets were all almost identical in terms of macronutrient makeup. It also is a bit less code to do the data from the whole world. Lmk if you guys wanna wanna broaden it up later if we aren't finding much difference between the EU and US on your side. I have some extra csv files in there I was thinking of incorporating but didn't want to overcomplicate it.